pyproject.toml
[project]
name = "jarvisfather"
version = "0.1.0"
description = "Multi-agent deployment SaaS via Telegram"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.13",
    "sqlalchemy>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "celery>=5.4",
    "redis>=5.2",
    "paramiko>=3.5",
    "cryptography>=44.0",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9",
    "httpx>=0.27",
    "sentence-transformers>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# torch CPU-версия (без CUDA, ~200MB вместо 2GB)
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
